videomixer: start/stop omx component
Start the omx component by changing the component
state from loaded to executing, allocate the buffers
to allow the transition from loaded to idle. The input
buffers are allocated by the mixer but it can be extended
to share the input buffers if they are omx buffers. The
output buffers for the first port will be allocated and
the other output ports will use the buffers allocated
for the first output port in order to have each of the
resultant images on the same output buffer.

Stop the omx component by changing the component
from executing to loaded, free the omx buffers to
allow the transition from idle to loaded.