omxvideomixer: send and get buffers from the mixer component.
Add a basic logic to handle buffers. Get the input
buffer for each of the sinkpads copy them and send them
to the omx mixer component. At starting poing, push the
output buffers to the mixer component to be used.
Get the processed buffers push them back to the omx
mixer component, at this point the element is not
outputing buffers to the downstream pipeline.