omxvideomixer: push an output buffer until the mixer component returned the corresponding omx buffer for each port
A mixer output buffer will be complete until each channel has
processed a corresponding buffer. So the element counts the
amount of buffers returned per buffer index and push them
to the the output queue until each of the ports returned
the buffer for an specific index. The output loop will
take care of getting the processed buffer from the queue,
create a gstreamer buffer with this data and push it
downstream.