omx: Deinitialize OMX
The plugin initialization calls OMX_Init() but OMX_Deinit() was
never called causing a memory leak each time an omx element is
used on a different process. Now OMX_Init will be called when
the first omx element is created and OMX_Deinit when the last
omx element is destroyed.