rrh264parser: wait for an IDR frame to generate the codec-data correctly
If the first buffer doesn't includes SPS and PPS, to generate
the codec data instead of assuming SPS and PPS values, the
rrh264parser will wait for an frame (IDR) to get the required
information. Also avoids the segmentation fault when no SPS
or PPS is found on the first buffer.